+++ src/pdocks.h
@@ -17,12 +17,12 @@ class PDOCKS {
     public:
         byte* finished;
         byte* used;
-        float* hittingNumArray;
-        float** D;
-        float** F;
+        double* hittingNumArray;
+        double** D;
+        double** F;
         int ALPHABET_SIZE;
-        float edgeCount;
-        float edgeNum;
+        double edgeCount;
+        double edgeNum;
         int k;
         int l;
         int total;
@@ -179,15 +179,15 @@ class PDOCKS {
         ofstream hittingStream(hittingPath);
         int hittingCount = 0;
         l = L-k+1;
-        hittingNumArray = new float[(int)edgeNum];
+        hittingNumArray = new double[(int)edgeNum];
         used = new byte[vertexExp];
         finished = new byte[vertexExp];
         topoSort = new byte[vertexExp];
-        D = new float*[l + 1];
-        float* Dpool = new float[(l+1)* vertexExp];
+        D = new double*[l + 1];
+        double* Dpool = new double[(l+1)* vertexExp];
         for(int i = 0; i < l+1; i++, Dpool += vertexExp) D[i] = Dpool;
-        F = new float*[l + 1];
-        float* Fpool = new float[(l+1)* vertexExp];
+        F = new double*[l + 1];
+        double* Fpool = new double[(l+1)* vertexExp];
         for(int i = 0; i < l+1; i++, Fpool += vertexExp) F[i] = Fpool;
         while (calculatePaths(l, threads)) {
             int imaxHittingNum = calculateHittingNumberParallel(l, false, threads);
@@ -208,7 +208,7 @@ Calculates hitting number for an edge of specified index with respect to a speci
 sequence length, counting paths of length L-k+1.
 @param i: Index of edge, L: Sequence length.
 */
-        float hittingNum = 0;
+        double hittingNum = 0;
         for (int j = (1 - edgeArray[i]) * L; j < L; j++) hittingNum = hittingNum + F[j][i % vertexExp] * D[(L-j-1)][i / ALPHABET_SIZE];
         hittingNumArray[i] = hittingNum;
     }
@@ -219,7 +219,7 @@ Calculates hitting number of all edges, counting paths of length L-k+1, in paral
 @return imaxHittingNum: Index of vertex with maximum hitting number.
 */  
         omp_set_dynamic(0);
-        float maxHittingNum = 0;
+        double maxHittingNum = 0;
         int imaxHittingNum = -1;
         #pragma omp parallel for num_threads(threads)
         for (int i = 0; i < (int)edgeNum; i++) calculateForEach(i, L);
